pyproject.toml
[project]
name = "ecsvt"
version = "0.1.0"
description = "Entity Resolution and CSV Transformation Tool - hybrid matching with fuzzy string matching, semantic embeddings, and blocking"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Redsky Analytics"}
]
keywords = [
    "entity-resolution",
    "duplicate-detection",
    "record-linkage",
    "fuzzy-matching",
    "semantic-similarity",
    "data-cleaning",
    "csv-transformation"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Information Analysis",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "faiss-cpu>=1.7.4",
    "ipykernel>=7.0.0",
    "jellyfish>=1.2.1",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.1",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.1",
]

[project.urls]
Homepage = "https://github.com/redsky-analytics/e-csvt"
Repository = "https://github.com/redsky-analytics/e-csvt"
Issues = "https://github.com/redsky-analytics/e-csvt/issues"

[project.scripts]
ecsvt-transform = "ecsvt.apply_transformations:main"
ecsvt-similarity = "ecsvt.cli.semantic_similarity_repl:run_repl"

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatchling.build.targets.wheel]
packages = ["src/ecsvt"]
